current-recording-list uses message IDs
Include a message ID in "save" and "upload-recording" messages and keep
track of the sent IDs. When a "save" or "upload-recording" response is
received, ignore the response if its ID is unknown. This prevents other
clients' actions from creating save/upload notifications for this client
instance, and prevents upload events from opening an unexpected Grafana
tab.