Add recording action for viewing in Grafana
Adds a recording action to the list row overflow action menu labelled
"View in Grafana ...". When the user clicks this, an "upload-recording"
command is sent to ContainerJFR with the recording name and Grafana
datasource URL as args (as is the upload-saved protocol currently). This
results in ContainerJFR POSTing the recording stream to the
jfr-datasource in-cluster and responding to the Web client when the
upload is complete. The Web client generates a message ID to include
with the original request, and when this ID is seen on a corresponding
upload-recording response, the Web client opens a new browser tab with
the Grafana dashboard URL configured on the ContainerJFR instance.

Closes #74