Connectionless API refactor
Switch to connectionless APIs. Remove connect/disconnect/is-connected commands
which explicitly set or get the state of the ContainerJFR<->Target connection,
which is no longer stateful.

For now, the client-side CommandChannel does maintain a piece of state
for the "selected" target, which is preserved only for purposes of
allowing the UI to mostly ignore the underlying change, while allowing
flexibility for future change.