View in Grafana (#83)
* Move recording and reports download actions into dropdown

* Use ApiService for recording and report downloads

ApiService includes credentials on requests and downloads files into
browser memory, then produces a virtual anchor link and "clicks" it,
allowing the request filename to also be modified. This is used for
downloading both JFR binary recordings as well as HTML reports for the
user to store locally (or for reports, to open in a new browser tab if
they wish)

* Add recording action for viewing in Grafana

Adds a recording action to the list row overflow action menu labelled
"View in Grafana ...". When the user clicks this, an "upload-recording"
command is sent to ContainerJFR with the recording name and Grafana
datasource URL as args (as is the upload-saved protocol currently). This
results in ContainerJFR POSTing the recording stream to the
jfr-datasource in-cluster and responding to the Web client when the
upload is complete. The Web client generates a message ID to include
with the original request, and when this ID is seen on a corresponding
upload-recording response, the Web client opens a new browser tab with
the Grafana dashboard URL configured on the ContainerJFR instance.

Closes #74