Remove timed background command channel ping
The periodic ping was initially used for development testing to check
whether communication errors were due to closed connections or the
backend not responding, but this is no longer a common concern and so
the ping serves no practical purpose. It should be removed to prevent it
from clogging backend logs.