Fix bug preventing archived recording uploads
A previous commit refactored the ApiService in a way that caused all
upload requests to use the HTTP API path as if they were all active
recordings. This re-adds a differentiation in the service implementation
between archived and active recordings to reflect the different HTTP API
paths for uploading either to a jfr-datasource.

Fixes #119