feat(archiveuploads): display prompt on cancel or navigation (#333)
Displays a React dialog prompt when the user clicks the Cancel button during an active upload, and displays the browser's native prompt to confirm when the user tries to refresh or close the tab/window. Also uses an AbortController/AbortSignal to actually instruct the browser to stop the file upload request if the user does confirm they wish to cancel the upload.